+++ machine_learning_module/c_example/e1.c
@@ -1,0 +1,5 @@
+#include <stdio.h>
+
+int main(){
+    printf("hello\n");
+}
